Fix: sitemap updating API
