+++ .github/workflows/deploy-frontend-dev.yml
@@ -1,8 +1,11 @@
 name: DeployFrontendDev
 
 on:
-  push:
-    branches:
+  workflow_run:
+    workflows: ["DeployBackendDev"]
+    types:
+      - completed
+    branches: 
       - "master"
 
 jobs:
