+++ nuxt-frontend/.prettierignore
@@ -1,0 +1,2 @@
+*.yml
+*.json
