+++ sitemap/sitemap_test.go
@@ -17,6 +17,7 @@ import (
 )
 
 //go:embed templates/career-email-template.html
+//go:embed templates/path.txt
 var templateFS embed.FS
 
 var repo *SitemapRepository
@@ -31,11 +32,10 @@ func TestInit(t *testing.T) {
 	}
 
 	careerRepo = jobs.New(testDB, templateFS, &stubUtilsRepo{})
-	repo = New(careerRepo)
+	repo = New(careerRepo, templateFS)
 }
 
 func TestSitemap(t *testing.T) {
-	testFolderPath = "../nuxt-frontend/pages"
 	engine := gin.New()
 	engine.GET("/api/sitemap", repo.GenerateSitemap)
 	req, err := http.NewRequest("GET", "/api/sitemap?baseUrl=http://localhost:8080", nil)
@@ -90,22 +90,24 @@ func expectedURLData() []URL {
 
 	sitemapUrls := []URL{
 		{Loc: baseUrl, Priority: `1`},
+		{Loc: baseUrl + "/services", Priority: `0.9`},
+		{Loc: baseUrl + "/portfolio", Priority: `0.9`},
+		{Loc: baseUrl + `/contributions`, Priority: `0.9`},
 		{Loc: baseUrl + "/resources", Priority: `0.9`},
 		{Loc: baseUrl + "/about", Priority: `0.9`},
-		{Loc: baseUrl + "/android-app-development", Priority: `0.9`},
 		{Loc: baseUrl + "/contact", Priority: `0.9`},
-		{Loc: baseUrl + `/contributions`, Priority: `0.9`},
+		{Loc: baseUrl + "/android-app-development", Priority: `0.9`},
 		{Loc: baseUrl + "/ios-app-development", Priority: `0.9`},
 		{Loc: baseUrl + `/mobile-app-development`, Priority: `0.9`},
-		{Loc: baseUrl + `/services`, Priority: `0.9`},
 		{Loc: baseUrl + `/thank-you`, Priority: `0.9`},
-		{Loc: baseUrl + "/portfolio", Priority: `0.9`},
-		{Loc: baseUrl + "/portfolio/luxeradio", Priority: `0.9`},
-		{Loc: baseUrl + "/portfolio/togness", Priority: `0.9`},
-		{Loc: baseUrl + "/portfolio/justly", Priority: `0.9`},
+		{Loc: baseUrl + `/jobs/thank-you`, Priority: `0.9`},
+		{Loc: baseUrl + `/unsubscribe`, Priority: `0.9`},
 		{Loc: baseUrl + "/jobs", Priority: `1`},
 		{Loc: baseUrl + "/jobs/ios-developer-a9b45f34-a1a5-419f-b536-b7c290925d6d", Priority: `0.9`},
 		{Loc: baseUrl + "/jobs/apply/ios-developer-a9b45f34-a1a5-419f-b536-b7c290925d6d", Priority: `0.9`},
+		{Loc: baseUrl + "/portfolio/luxeradio", Priority: `0.9`},
+		{Loc: baseUrl + "/portfolio/togness", Priority: `0.9`},
+		{Loc: baseUrl + "/portfolio/justly", Priority: `0.9`},
 	}
 
 	for i := range sitemapUrls {
