+++ templates/path.txt
@@ -1,0 +1,12 @@
+services
+portfolio
+contributions
+resources
+about
+contact
+android-app-development
+ios-app-development
+mobile-app-development
+thank-you
+jobs/thank-you
+unsubscribe
